provider/core: changed name from serialze application to save snapshot
